Added Firebase accounts to Login and Signup Activities
Firebase is now linked to the app and you can make a user account.

If you are not logged in, the main activity will automatically take you
to the login screen. You can login or create a new account. When you
login or create a new account, you will be taken to the MainActivity
again, until you sign out.